\name{te-package}
\alias{tme-package}
\alias{tme}
\docType{package}
\title{Teaching Marketing Engineering with R}
\description{
This package provides 
}
\details{
\tabular{ll}{
Package: \tab tme\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2006-12-19\cr
License: \tab GPL version 2.1 or newer\cr
}

}
\author{
  Christian Buchta
  Stefan Theussl

Maintainer: Stefan Theussl <stefan.theussl@wu-wien.ac.at>
}
\references{
Marketing Engineering
}
\keyword{ package }
\seealso{
  %\code{\link[lpSolve:lpSolve-package]{lpSolve}}
}
\examples{
}
